Refactored pages/index.js to now export src/index.js which now uses higher order components for the Score, AddQuestion, and QuestionList components.
